requirements.txt
Django==3.1.*
channels==2.4.0
channels_redis==3.*
django-widget-tweaks==1.4.*
django-environ==0.4.*
django-extensions==2.1.*
djangorestframework==3.11.1
Pillow==6.*
psycopg2==2.*
uvicorn
